requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.10.0
Django==1.10
django-extensions==1.7.9
django-model-utils==3.0.0
packaging==16.8
pycparser==2.17
pyparsing==2.2.0
six==1.10.0
